backend/requirements.txt
aiofiles==0.5.0
alembic==1.4.2
click==7.1.2
fastapi==0.59.0
h11==0.9.0
httptools==0.1.1
Mako==1.1.3
MarkupSafe==1.1.1
py-bcrypt==0.4
pydantic==1.6
PyJWT==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.18
starlette==0.13.4
uvicorn==0.11.5
uvloop==0.14.0
websockets==8.1
